siraj_ai_service/requirements.txt
flask
requests
openai-whisper
torch
